requirements.txt
BeautifulSoup
Django==1.6.1
Markdown
MySQL-python
South
Werkzeug
argparse
django-debug-toolbar
django-extensions
django-crispy-forms
django-picklefield
django-tracking2
django-bootstrap3
django-mobile
gunicorn
slumber
ipython
jsonfield
requests
simplejson
six
sqlparse
uWSGI
wsgiref
redis
django-daterange-filter
django-dashing
django-social-auth
requests-toolbelt
django-js-reverse
requests-futures
dj_database_url
ssh
psycopg2
django_markdown
djangorestframework
django-filter